+++ src/main/java/com/dearjolly/server/domain/feedback/entity/Feedbacks.java
@@ -1,6 +1,6 @@
-package com.dearjolly.server.entity;
+package com.dearjolly.server.domain.feedback.entity;
 
-import com.dearjolly.server.entity.enums.Status;
+import com.dearjolly.server.domain.letter.entity.Letters;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -14,7 +14,6 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
