+++ src/main/java/com/dearjolly/server/domain/feedback/enums/CorrectionType.java
@@ -1,4 +1,4 @@
-package com.dearjolly.server.entity.enums;
+package com.dearjolly.server.domain.feedback.enums;
 
 public enum CorrectionType {
     UNCHANGED,
