+++ src/main/java/com/dearjolly/server/domain/letter/enums/Status.java
@@ -1,4 +1,4 @@
-package com.dearjolly.server.entity.enums;
+package com.dearjolly.server.domain.letter.enums;
 
 public enum Status {
     DRAFT("작성중"),
