+++ src/main/java/com/dearjolly/server/domain/user/entity/Users.java
@@ -1,7 +1,7 @@
-package com.dearjolly.server.entity;
+package com.dearjolly.server.domain.user.entity;
 
-import com.dearjolly.server.entity.enums.Role;
-import com.dearjolly.server.entity.enums.Status;
+import com.dearjolly.server.domain.user.enums.Role;
+import com.dearjolly.server.domain.letter.entity.Letters;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -13,7 +13,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
